kselftest/arm64: Add 2024 dpISA extensions to hwcap test
commit 8600640d21cf90f3c5c4f06a5b214fbe4be9a74a upstream.

Add coverage of the hwcaps for the 2024 dpISA extensions to the hwcap
test.

We don't actually test SIGILL generation for CMPBR since the need to
branch makes it a pain to generate and the SIGILL detection would be
unreliable anyway. Since this should be very unusual we provide a stub
function rather than supporting a missing test.

The sigill functions aren't well sorted in the file so the ordering is a
bit random.

Signed-off-by: Mark Brown <broonie@kernel.org>
Link: https://lore.kernel.org/r/20250107-arm64-2024-dpisa-v5-5-7578da51fc3d@kernel.org
Signed-off-by: Will Deacon <will@kernel.org>
[ bh: backport to v6.6 - re-anchor the arm64 HWCAP selftest to the target test layout. ]
Signed-off-by: Bin Huang <bin.huang2@arm.com>